implement subtract

+++ math.h
@@ -1,4 +1,4 @@
-double add(int n1, int n2);
-double subtract(double n1, double n2);
-double multiply(double n1, double n2);
-double divide(double n1, double n2);
+int add(int n1, int n2);
+int subtract(int n1, int n2);
+int multiply(int n1, int n2);
+int divide(int n1, int n2);
